requirements.txt
# Production dependencies
fastapi>=0.104.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
click>=8.1.0
